+++ repos/base-codezero/src/core/include/ipc_pager.h
@@ -71,7 +71,7 @@ namespace Genode {
 	/**
 	 * Special paging server class
 	 */
-	class Ipc_pager : public Native_capability
+	class Ipc_pager
 	{
 		private:
 
@@ -95,11 +95,6 @@ namespace Genode {
 //			void _reply_and_wait();
 
 		public:
-
-			/**
-			 * Constructor
-			 */
-			Ipc_pager();
 
 			/**
 			 * Wait for a new page fault received as short message IPC
@@ -141,9 +136,9 @@ namespace Genode {
 			void acknowledge_wakeup();
 
 			/**
-			 * Return thread ID of last faulter
+			 * Returns true if the last request was send from a core thread
 			 */
-			Native_thread_id last() const { return _last; }
+			bool request_from_core() { return true; }
 
 			/**
 			 * Return badge for faulting thread
