+++ repos/base-codezero/src/core/pager.cc
@@ -1,74 +1,174 @@
 /*
- * \brief  Dummy pager framework
+ * \brief  Pager support for Codezero
  * \author Norman Feske
- * \date   2009-10-02
+ * \date   2010-02-16
  */
 
 /*
- * Copyright (C) 2009-2013 Genode Labs GmbH
+ * Copyright (C) 2010-2013 Genode Labs GmbH
  *
  * This file is part of the Genode OS framework, which is distributed
  * under the terms of the GNU General Public License version 2.
  */
 
-/* Core includes */
+/* Genode includes */
+#include <base/printf.h>
 #include <pager.h>
 
+/* Codezero includes */
+#include <ipc_pager.h>
+#include <codezero/syscalls.h>
+
+
 using namespace Genode;
+using namespace Codezero;
+
+enum { verbose_page_faults = false };
 
 
-/**********************
- ** Pager activation **
- **********************/
+/************************
+ ** Page-fault utility **
+ ************************/
 
-void Pager_activation_base::entry()
+class Fault
 {
-	Ipc_pager pager;
-	_cap = pager;
-	_cap_valid.unlock();
+	public:
 
-	bool reply = false;
+		enum Type { READ, WRITE, EXEC, UNKNOWN };
 
-	while (1) {
+	private:
 
-		if (reply)
-			pager.reply_and_wait_for_fault();
-		else
-			pager.wait_for_fault();
+		/**
+		 * Translate Codezero page-fault information to generic fault type
+		 *
+		 * \param sr status
+		 * \param pte page-table entry
+		 */
+		static Type _fault_type(umword_t sr, umword_t pte)
+		{
+			if (is_prefetch_abort(sr))
+				return EXEC;
 
-		/* lookup referenced object */
-		Object_pool<Pager_object>::Guard _obj(_ep ? _ep->lookup_and_lock(pager.badge()) : 0);
-		Pager_object * obj = _obj;
-		reply = false;
+			if ((pte & PTE_PROT_MASK) == (__MAP_USR_RO & PTE_PROT_MASK))
+				return WRITE;
 
-		/* handle request */
-		if (obj) {
-			reply = !obj->pager(pager);
-			/* something strange occurred - leave thread in pagefault */
+			return READ;
+		}
+
+		Type     _type;
+		umword_t _addr;
+		umword_t _ip;
+
+	public:
+
+		/**
+		 * Constructor
+		 *
+		 * \param kdata  Codezero-specific page-fault information
+		 */
+		Fault(struct fault_kdata const &kdata)
+		:
+			_type(_fault_type(kdata.fsr, kdata.pte)),
+			_addr(_type == EXEC ? kdata.faulty_pc : kdata.far),
+			_ip(kdata.faulty_pc)
+		{ }
+
+		Type     type() const { return _type; }
+		umword_t addr() const { return _addr; }
+		umword_t ip()   const { return _ip; }
+};
+
+
+/**
+ * Print page-fault information in a human-readable form
+ */
+inline void print_page_fault(Fault &fault, int from)
+{
+	printf("page (%s%s%s) fault from %d at pf_addr=%lx, pf_ip=%lx\n",
+	       fault.type() == Fault::READ  ? "r"  : "-",
+	       fault.type() == Fault::WRITE ? "w"  : "-",
+	       fault.type() == Fault::EXEC  ? "x"  : "-",
+	       from, fault.addr(), fault.ip());
+}
+
+
+/***************
+ ** IPC pager **
+ ***************/
+
+void Ipc_pager::wait_for_fault()
+{
+	for (;;) {
+		int ret = l4_receive(L4_ANYTHREAD);
+
+		if (ret < 0) {
+			PERR("pager: l4_received returned ret=%d", ret);
 			continue;
 		}
 
-		/*
-		 * We got a request from one of cores region-manager sessions
-		 * to answer the pending page fault of a resolved region-manager
-		 * client. Hence, we have to send the page-fault reply to the
-		 * specified thread and answer the call of the region-manager
-		 * session.
-		 *
-		 * When called from a region-manager session, we receive the
-		 * core-local address of the targeted pager object via the
-		 * first message word, which corresponds to the 'fault_ip'
-		 * argument of normal page-fault messages.
-		 */
-		obj = reinterpret_cast<Pager_object *>(pager.fault_ip());
+		umword_t tag         = l4_get_tag();
+		int      faulter_tid = l4_get_sender();
 
-		/* send reply to the calling region-manager session */
-		pager.acknowledge_wakeup();
+		if (tag != L4_IPC_TAG_PFAULT) {
+			PWRN("got an unexpected IPC from %d", faulter_tid);
+			continue;
+		}
 
-		/* answer page fault of resolved pager object */
-		pager.set_reply_dst(obj->cap());
-		pager.acknowledge_wakeup();
+		/* copy fault information from message registers */
+		struct fault_kdata fault_kdata;
+		for (unsigned i = 0; i < sizeof(fault_kdata_t)/sizeof(umword_t); i++)
+			((umword_t *)&fault_kdata)[i] = read_mr(MR_UNUSED_START + i);
+
+		Fault fault(fault_kdata);
+
+		if (verbose_page_faults)
+			print_page_fault(fault, faulter_tid);
+
+		/* determine corresponding page in our own address space */
+		_pf_addr  = fault.addr();
+		_pf_write = fault.type() == Fault::WRITE;
+		_pf_ip    = fault.ip();
+		_last     = faulter_tid;
+
+		return;
 	}
+}
+
+
+void Ipc_pager::reply_and_wait_for_fault()
+{
+	/* install mapping */
+	umword_t flags = _reply_mapping.writeable() ? MAP_USR_RW
+	                                            : MAP_USR_RO;
+	
+	/*
+	 * XXX: remove heuristics for mapping device registers.
+	 */
+	if (_reply_mapping.from_phys() == 0x10120000  /* LCD */
+	 || _reply_mapping.from_phys() == 0x10006000  /* keyboard */
+	 || _reply_mapping.from_phys() == 0x10007000) /* mouse */
+		flags = MAP_USR_IO;
+
+	int ret = l4_map((void *)_reply_mapping.from_phys(),
+	                 (void *)_reply_mapping.to_virt(),
+	                 _reply_mapping.num_pages(), flags, _last);
+
+	/* wake up faulter if mapping succeeded */
+	if (ret < 0)
+		PERR("l4_map returned %d, putting thread %d to sleep", ret, _last);
+	else
+		acknowledge_wakeup();
+
+	/* wait for next page fault */
+	wait_for_fault();
+}
+
+
+void Ipc_pager::acknowledge_wakeup()
+{
+	enum { SUCCESS = 0 };
+	l4_set_sender(_last);
+	l4_ipc_return(SUCCESS);
 }
 
 
@@ -76,30 +176,7 @@ void Pager_activation_base::entry()
  ** Pager entrypoint **
  **********************/
 
-Pager_entrypoint::Pager_entrypoint(Cap_session *, Pager_activation_base *a)
-: _activation(a)
-{ _activation->ep(this); }
-
-
-void Pager_entrypoint::dissolve(Pager_object *obj)
+Untyped_capability Pager_entrypoint::_manage(Pager_object *obj)
 {
-	remove_locked(obj);
-}
-
-
-Pager_capability Pager_entrypoint::manage(Pager_object *obj)
-{
-	/* return invalid capability if no activation is present */
-	if (!_activation) return Pager_capability();
-
-	_activation->cap();
-
-	Untyped_capability cap = Native_capability(_activation->cap().dst(), obj->badge());
-
-	/* add server object to object pool */
-	obj->cap(cap);
-	insert(obj);
-
-	/* return capability that uses the object id as badge */
-	return reinterpret_cap_cast<Pager_object>(cap);
+	return Untyped_capability(_tid.l4id, obj->badge());
 }
