+++ repos/base-codezero/src/core/target.inc
@@ -12,11 +12,11 @@ SRC_CC      += cap_session_component.cc \
                dump_alloc.cc \
                io_mem_session_component.cc \
                io_mem_session_support.cc \
-               ipc_pager.cc \
                irq_session_component.cc \
                main.cc \
                pager.cc \
-               pager_common.cc \
+               pager_ep.cc \
+               pager_object.cc \
                pd_session_component.cc \
                platform.cc \
                platform_pd.cc \
@@ -58,5 +58,6 @@ vpath dataspace_component.cc      $(GEN_CORE_DIR)
 vpath core_mem_alloc.cc           $(GEN_CORE_DIR)
 vpath dump_alloc.cc               $(GEN_CORE_DIR)
 vpath context_area.cc             $(GEN_CORE_DIR)
-vpath pager_common.cc             $(GEN_CORE_DIR)
+vpath pager_ep.cc                 $(GEN_CORE_DIR)
+vpath pager_object.cc             $(GEN_CORE_DIR)
 vpath %.cc                        $(REP_DIR)/src/core
