+++ repos/base-fiasco/src/core/target.inc
@@ -12,12 +12,12 @@ SRC_CC      += cap_session_component.cc \
                io_mem_session_component.cc \
                io_mem_session_support.cc \
                io_port_session_component.cc \
-               ipc_pager.cc \
                irq_session_component.cc \
                main.cc \
                multiboot_info.cc \
                pager.cc \
-               pager_common.cc \
+               pager_ep.cc \
+               pager_object.cc \
                pd_session_component.cc \
                platform.cc \
                platform_pd.cc \
@@ -58,6 +58,7 @@ vpath trace_session_component.cc  $(GEN_CORE_DIR)
 vpath dataspace_component.cc      $(GEN_CORE_DIR)
 vpath dump_alloc.cc               $(GEN_CORE_DIR)
 vpath context_area.cc             $(GEN_CORE_DIR)
-vpath pager_common.cc             $(GEN_CORE_DIR)
+vpath pager_object.cc             $(GEN_CORE_DIR)
+vpath pager_ep.cc                 $(GEN_CORE_DIR)
 vpath core_printf.cc              $(BASE_DIR)/src/base/console
 vpath %.cc                        $(REP_DIR)/src/core
