+++ repos/base-sel4/src/core/pager.cc
@@ -97,51 +97,7 @@ void Ipc_pager::acknowledge_wakeup()
 }
 
 
-Ipc_pager::Ipc_pager()
-:
-	Native_capability(Capability_space::create_ep_cap(*Thread_base::myself())),
-	_last(0)
-{ }
-
-
-
-/**********************
- ** Pager activation **
- **********************/
-
-void Pager_activation_base::entry()
-{
-	Ipc_pager pager;
-	_cap = pager;
-	_cap_valid.unlock();
-
-	bool reply_pending = false;
-	while (1) {
-
-		if (reply_pending)
-			pager.reply_and_wait_for_fault();
-		else
-			pager.wait_for_fault();
-
-		reply_pending = false;
-
-		/* lookup referenced object */
-		Object_pool<Pager_object>::Guard _obj(_ep ? _ep->lookup_and_lock(pager.badge()) : 0);
-		Pager_object *obj = _obj;
-
-		/* handle request */
-		if (obj) {
-			if (pager.is_exception()) {
-				obj->submit_exception_signal();
-				continue;
-			}
-
-			/* send reply if page-fault handling succeeded */
-			if (!obj->pager(pager))
-				reply_pending = true;
-		}
-	}
-}
+Ipc_pager::Ipc_pager() : _last(0) { }
 
 
 /******************
@@ -164,39 +120,16 @@ void Pager_object::unresolved_page_fault_occurred()
  ** Pager entrypoint **
  **********************/
 
-Pager_entrypoint::Pager_entrypoint(Cap_session *, Pager_activation_base *a)
-: _activation(a)
-{ _activation->ep(this); }
-
-
-void Pager_entrypoint::dissolve(Pager_object *obj)
+Untyped_capability Pager_entrypoint::_manage(Pager_object *obj)
 {
-	remove_locked(obj);
-}
-
-
-Pager_capability Pager_entrypoint::manage(Pager_object *obj)
-{
-	/* return invalid capability if no activation is present */
-	if (!_activation) return Pager_capability();
-
 	/*
 	 * Create minted endpoint capability of the pager entrypoint.
 	 * The badge of the page-fault message is used to find the pager
 	 * object for faulted thread.
 	 */
-	Native_capability ep_cap = _activation->cap();
-
 	Rpc_obj_key rpc_obj_key((addr_t)obj->badge());
 
-	Untyped_capability new_obj_cap =
-		Capability_space::create_rpc_obj_cap(ep_cap, 0, rpc_obj_key);
-
-	/* add server object to object pool */
-	obj->cap(new_obj_cap);
-	insert(obj);
-
-	return reinterpret_cap_cast<Pager_object>(new_obj_cap);
+	Untyped_capability ep_cap(Capability_space::create_ep_cap(*this));
+	return Capability_space::create_rpc_obj_cap(ep_cap, nullptr, rpc_obj_key);
 }
-
